url_shortener/requirements.txt
anyio==3.6.1
black==22.6.0
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
ecdsa==0.18.0
fastapi==0.79.0
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
py-bcrypt==0.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
validators==0.20.0
